Update launcher icons generator with elevation support
Utilizing a modified version of the StackBlur algorithm created by Mario
Klingemann (see
http://www.quasimondo.com/StackBlurForCanvas/StackBlurDemo.html) that
will play nicely with the rest of the app, the elevation is created by
setting the color of the top layer to a semi-transparent black and then
blurring it with a ~3dp radius Gaussian blur to create the background.
Additionally, the shadow for the long shadow has been set to an alpha
value of 50, an opacity of 20%, as the material design spec indicates
for drop shadows.