Update launcher icons with shadow intensity changing
It can be hard to see the shadow generated for the foreground shapes for
darker colors, and also on light colors the shadow might be too dark.
The use of a slider allows for the user to modify the shadow intensity
in these situations and modify how dark it is. The shadow intensity can
now be changed for the drop-shadow (elevation), the long-shadow, and the
score effects.